Skaffold for local deployment (#93)
* Change extension controller pod anti affinity to preferred (#84)

* Change extension controller pod anti affinity to preferred

* Fix label value in podAntiAffinity

Co-authored-by: Dimitar Mirchev <dimitar.mirchev@sap.com>

---------

Co-authored-by: Dimitar Mirchev <dimitar.mirchev@sap.com>

* Build and deploy the apps in the right order, patch imaveVectorOverwrite

A post-build skaffold hook is used for patching the imageVectorOverwrite
field of the controllerddeployment.

This way, the lakom admission controller is first built and its name is
patched in using a skaffold hook + kustomize.

* Add cosign public key to kustomize deploy step

* Extract image repo&tag from skaffold IMAGE var

For some reason, SKAFFOLD_IMAGE_REPO does not contain the substring
before the last ':' but instead contains the substring before the first
'/'. This means that if
SKAFFOLD_IMAGE=localhost:5001/local-skaffold_lakom:sometag
then
SKAFFOLD_IAMGE_TAG=sometag
SKAFFOLD_IMAGE_REPO=localhost:5001
instead of 'localhost:5001/local-skaffold_lakom

* Add the imagevector-overwrite patch to the example/kustomization.yaml file

* Fix expected structure of controller-registration.yaml

* Fix cosign pub key & add skaffold env var comments

* Address PR comments

* Add crane as dep and manually append image digest in hook

* Use gardener skaffold script instead of local

* Use crane for extension-up

* Address PR comments

* Remove crane dependancy and use garden.local.gardener.cloud

Now the lakom pods can reach the in-cluster registry via
`garden.local.gardener.cloud` instead of localhost. This removes the
need for adding the digest manually to the end of the image tag which we
needed crane for.

* Add documentation for developing locally with skaffold

* Address PR comments

* Add feature for allowing HTTP communication with registries

* Tag images in skaffold using digest

This way deployments will detect the new image name and restart pods
with the new image version.

* Allow insecure communication for verifier

* Add insecure option to resolver type instead of changing the Verifier interface

* Add forgotten backslash to makefile

---------

Co-authored-by: Vladimir Nachev <vladimir.nachev@sap.com>
Co-authored-by: Dimitar Mirchev <dimitar.mirchev@sap.com>