Extract rules to the helm chart values
In the Gardener context, multiple clusters exist that server different
purposes and contain different types of resources.
Lakom is intended to verify all resources that contain some sort of
reference to an OCI artifact. However, we don't want to install the same
validating and mutating webhook configs to all the clusters due to the
mentioned differences.

The webhook rules can now be configured when installing the lakom helm
chart.This allows us to be more flexible on the resources that lakom
will validate.